Restructured tests module to match django_gvar struct
