+++ tests/field_tests/tests/__init__.py
@@ -1,0 +1,4 @@
+"""Submodule provides class and tests for GVar fields."""
+from field_tests.tests.models_tests import *
+from field_tests.tests.forms_tests import *
+from field_tests.tests.parser_tests import *
