+++ tests/field_tests/tests/forms_tests.py
@@ -1,0 +1,8 @@
+"""Tests for GVar field."""
+from django.test import TestCase
+
+from gvar import gvar
+
+from django_gvar.testing import assert_allclose
+
+from field_tests.models import ExampleTable
