+++ tests/pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = django_gvar_tests.settings
-python_files = tests.py test_*.py *_tests.py
+python_files = *_tests.py
 addopts = -v
